+++ src/components/CardBtn/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+import "./style.css";
+
+function CardBtn(props) {
+  return (
+    <button 
+      onClick={props.onClick} 
+      className={`card-btn ${props["data-value"]}`} 
+      data-value={props["data-value"]} 
+    />
+  );
+}
+
+export default CardBtn;
