+++ src/components/CardBtn/style.css
@@ -1,0 +1,50 @@
+.card-btn {
+  border-radius: 50%;
+  height: 50px;
+  width: 50px;
+  position: absolute;
+  border: none;
+  bottom: 10%;
+  color: #FFF;
+  font-size: 24px;
+  cursor: pointer;
+  font-family: FontAwesome;
+  outline: none;
+}
+
+.pass {
+  left: 1%;
+  background: #E53935;
+}
+
+.pass::after {
+  content: "\f165";
+}
+
+.pick {
+  right: 1%;
+  background: #43A047;
+}
+
+.pick::after {
+  content: "\f164";
+}
+
+.pick:hover::after, .pass:hover::after {
+  animation: iconBounce .57s 1;
+}
+
+@keyframes iconBounce {
+  0% {
+    font-size: 24px;
+  }
+  
+   50% {
+    font-size: 28px;
+  }
+  
+  100% {
+    font-size: 24px;
+  }
+
+}
