+++ src/components/Footer/index.js
@@ -4,7 +4,7 @@ import "./style.css";
 function Footer() {
   return (
     <footer className="footer">
-      <span>Solemate 2020</span>
+      <span>Pupster 2017</span>
     </footer>
   );
 }
