--- src/components/Home/index.js
@@ -1,2 +1,0 @@
-import react from React;
-
