+++ src/components/Wrapper/style.css
@@ -1,4 +1,4 @@
 .wrapper {
-    width:100%;
-    padding:10px;
+  min-height: 100%;
+  padding-bottom: 100px;
 }
