+++ src/pages/Discover.js
@@ -1,8 +1,7 @@
 import React, { Component } from "react";
 import API from "../utils/API";
-// import Users from "../users.json";
 import Card from "../components/Card";
-import Alert from "../components/Alert";
+
 
 class Discover extends Component {
   state = {
@@ -55,15 +54,11 @@ class Discover extends Component {
       <div>
         <h1 className="text-center">Make New Friends</h1>
         <h3 className="text-center">
-          Thumbs up you'd like to meet!
+          Meet New People
         </h3>
         <Card image={this.state.image} handleBtnClick={this.handleBtnClick} />
-        <h1 className="text-center">
-          Made friends with {this.state.matchCount} so far!
-        </h1>
-        <Alert style={{ opacity: this.state.match ? 1 : 0 }} type="success">
-          Yay!
-        </Alert>
+        
+        
       </div>
     );
   }
