+++ src/utils/API.js
@@ -7,4 +7,3 @@ export default {
     return axios.get("https://randomuser.me/api/?results=5000");
   }
 };
-
